feat(structure): reorganización del proyecto para implementar Bounded Contexts según DDD
