+++ __tests__/domain/podcastDetailsTransformer.test.ts
@@ -1,4 +1,4 @@
-import { transformPodcastDetails } from '../../src/domain/services/podcastDetailsTransformer';
+import { transformPodcastDetails } from '../../src/podcastManagement/domain/services/podcastDetailsTransformer';
 
 describe('Podcast Details Transformer', () => {
   it('Debería transformar datos válidos de la API en detalles de podcast y episodios', () => {
