+++ __tests__/ui/HomePage.test.tsx
@@ -3,13 +3,13 @@ import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { render, screen } from '@testing-library/react';
 import { act } from 'react-dom/test-utils';
-import HomePage from './../../src/ui/pages/HomePage';
-import { podcastService } from './../../src/infrastructure/repositories/podcastService';
+import HomePage from './../../src/podcastManagement/ui/pages/HomePage';
+import { podcastService } from './../../src/podcastManagement/infrastructure/repositories/podcastService';
 import { LoadingProvider } from './../../src/shared/context/LoadingContext';
-import { NavigationProvider } from './../../src/application/context/NavigationContext';
-import { PodcastServiceProvider } from './../../src/infrastructure/context/PodcastServiceContext';
+import { NavigationProvider } from './../../src/podcastManagement/application/context/NavigationContext';
+import { PodcastServiceProvider } from './../../src/podcastManagement/infrastructure/context/PodcastServiceContext';
 
-jest.mock('./../../src/infrastructure/repositories/podcastService');
+jest.mock('./../../src/podcastManagement/infrastructure/repositories/podcastService');
 
 describe('HomePage Component', () => {
   const mockPodcasts = [
