+++ src/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import HomePage from './ui/pages/HomePage';
-import PodcastDetailPage from './ui/pages/PodcastDetailPage';
-import EpisodeDetailPage from './ui/pages/EpisodeDetailPage';
-import Header from './ui/components/Header';
+import HomePage from './podcastManagement/ui/pages/HomePage';
+import PodcastDetailPage from './podcastManagement/ui/pages/PodcastDetailPage';
+import EpisodeDetailPage from './podcastManagement/ui/pages/EpisodeDetailPage';
+import Header from './podcastManagement/ui/components/Header';
 import { LoadingProvider } from './shared/context/LoadingContext';
-import { NavigationProvider } from './application/context/NavigationContext';
-import { PodcastServiceProvider } from './infrastructure/context/PodcastServiceContext';
+import { NavigationProvider } from './podcastManagement/application/context/NavigationContext';
+import { PodcastServiceProvider } from './podcastManagement/infrastructure/context/PodcastServiceContext';
 
 const App: React.FC = () => {
   return (
