+++ src/podcastManagement/domain/entities/podcastDetail.ts
